Conflicts:
	NMBS/bin/dao/AdresDAO.class
	NMBS/bin/dao/MedewerkerDAO.class
	NMBS/bin/gui/AdminGui$1.class
	NMBS/bin/gui/AdminGui$MenuItemHandler.class
	NMBS/bin/gui/AdminGui.class
	NMBS/bin/gui/GebruikerToevoegenGui$MenuItemHandler.class
	NMBS/bin/gui/GebruikerToevoegenGui.class
	NMBS/bin/source/Adres.class
	NMBS/bin/source/Klant.class
	NMBS/src/gui/AdminGui.java
	NMBS/src/gui/GebruikerToevoegenGui.java
	NMBS/src/source/Adres.java
	NMBS/src/source/Klant.java

+++ NMBS/src/gui/GebruikerBewerkenGui.java
@@ -12,8 +12,11 @@ import javax.swing.JFrame;
 import javax.swing.UIManager;
 import java.awt.Color;
 import com.jgoodies.forms.factories.DefaultComponentFactory;
+import com.sun.xml.internal.ws.policy.privateutil.PolicyUtils.Collections;
 
+import dao.LoginDao;
 import dao.MedewerkerDAO;
+import handler.Controller;
 import source.Adres;
 import source.Login;
 import source.Medewerker;
@@ -23,12 +26,12 @@ import source.Rol;
 import java.awt.Font;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.util.ArrayList;
-import java.util.Vector;
+import java.util.*;
 
 import javax.swing.JList;
 import javax.swing.JOptionPane;
 import javax.swing.JTextPane;
+
 
 public class GebruikerBewerkenGui extends JPanel {
 	private JTextField textField;
@@ -38,11 +41,14 @@ public class GebruikerBewerkenGui extends JPanel {
 	private ArrayList<Medewerker> arrayLijst;
 	private ArrayList<Object> objecten;
 	private JButton btnVerwijderen;
+	private JButton btnPasswordReset;
+	public String navigation;
+	
 	
 	public GebruikerBewerkenGui() {
 		setBackground(UIManager.getColor("CheckBoxMenuItem.selectionBackground"));
 		
-		JLabel lblGebruikerBewerken = DefaultComponentFactory.getInstance().createTitle("Gebruiker bewerken");
+		JLabel lblGebruikerBewerken = DefaultComponentFactory.getInstance().createTitle("Gebruikers beheren");
 		lblGebruikerBewerken.setFont(new Font("Tahoma", Font.PLAIN, 14));
 		
 		
@@ -51,22 +57,38 @@ public class GebruikerBewerkenGui extends JPanel {
 		arrayLijst = MedewerkerDAO.getAllMedewerkers();
 		
 		
+		
 		DefaultListModel<Medewerker> dlm = new DefaultListModel<Medewerker>();
+		
+		
 		for(Medewerker m : arrayLijst)
 		{
 			dlm.addElement(m);
 		}
+		
 		list = new JList<Medewerker>(dlm);
 		
 		textField = new JTextField();
 		textField.setColumns(10);
 		
 		btnZoeken = new JButton("Zoeken");
+		btnZoeken.setFont(new Font("Segoe UI", Font.BOLD, 14));
+		btnZoeken.setBackground(Color.ORANGE);
+		
 		btnBewerken = new JButton("Bewerken");
+		btnBewerken.setFont(new Font("Segoe UI", Font.BOLD, 14));
+		btnBewerken.setBackground(Color.ORANGE);
 		btnBewerken.addActionListener(new MenuItemHandler());
 		
 		btnVerwijderen = new JButton("Verwijderen");
+		btnVerwijderen.setFont(new Font("Segoe UI", Font.BOLD, 14));
+		btnVerwijderen.setBackground(Color.ORANGE);
 		btnVerwijderen.addActionListener(new MenuItemHandler());
+		
+		btnPasswordReset = new JButton("Password Reset");
+		btnPasswordReset.setFont(new Font("Segoe UI", Font.BOLD, 14));
+		btnPasswordReset.setBackground(Color.ORANGE);
+		btnPasswordReset.addActionListener(new MenuItemHandler());
 		
 		GroupLayout groupLayout = new GroupLayout(this);
 		groupLayout.setHorizontalGroup(
@@ -74,35 +96,41 @@ public class GebruikerBewerkenGui extends JPanel {
 				.addGroup(groupLayout.createSequentialGroup()
 					.addGap(37)
 					.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
-						.addGroup(groupLayout.createSequentialGroup()
-							.addComponent(list, GroupLayout.PREFERRED_SIZE, 197, GroupLayout.PREFERRED_SIZE)
-							.addGap(30)
-							.addGroup(groupLayout.createParallelGroup(Alignment.LEADING, false)
-								.addComponent(btnBewerken, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-								.addComponent(btnVerwijderen, GroupLayout.DEFAULT_SIZE, 101, Short.MAX_VALUE)))
+						.addComponent(lblGebruikerBewerken)
 						.addGroup(groupLayout.createSequentialGroup()
 							.addComponent(btnZoeken)
-							.addGap(18)
-							.addComponent(textField, GroupLayout.PREFERRED_SIZE, 112, GroupLayout.PREFERRED_SIZE))
-						.addComponent(lblGebruikerBewerken))
-					.addContainerGap(88, Short.MAX_VALUE))
+							.addPreferredGap(ComponentPlacement.RELATED, 90, Short.MAX_VALUE)
+							.addComponent(textField, GroupLayout.PREFERRED_SIZE, 110, GroupLayout.PREFERRED_SIZE))
+						.addGroup(groupLayout.createSequentialGroup()
+							.addComponent(list, GroupLayout.DEFAULT_SIZE, 283, Short.MAX_VALUE)
+							.addPreferredGap(ComponentPlacement.UNRELATED)))
+					.addGap(10)
+					.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
+						.addComponent(btnVerwijderen, GroupLayout.DEFAULT_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+						.addComponent(btnBewerken, GroupLayout.DEFAULT_SIZE, 113, Short.MAX_VALUE)
+						.addComponent(btnPasswordReset, GroupLayout.PREFERRED_SIZE, 113, Short.MAX_VALUE))
+					.addContainerGap())
 		);
 		groupLayout.setVerticalGroup(
 			groupLayout.createParallelGroup(Alignment.LEADING)
 				.addGroup(groupLayout.createSequentialGroup()
-					.addGap(12)
-					.addComponent(lblGebruikerBewerken)
-					.addGap(27)
-					.addGroup(groupLayout.createParallelGroup(Alignment.BASELINE)
-						.addComponent(btnZoeken)
-						.addComponent(textField, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE))
-					.addPreferredGap(ComponentPlacement.UNRELATED)
 					.addGroup(groupLayout.createParallelGroup(Alignment.LEADING)
 						.addGroup(groupLayout.createSequentialGroup()
+							.addGap(12)
+							.addComponent(lblGebruikerBewerken)
+							.addGap(27)
+							.addGroup(groupLayout.createParallelGroup(Alignment.BASELINE)
+								.addComponent(btnZoeken)
+								.addComponent(textField, GroupLayout.PREFERRED_SIZE, 27, GroupLayout.PREFERRED_SIZE))
+							.addGap(12)
+							.addComponent(list, GroupLayout.DEFAULT_SIZE, 228, Short.MAX_VALUE))
+						.addGroup(groupLayout.createSequentialGroup()
+							.addGap(97)
 							.addComponent(btnBewerken)
-							.addGap(18)
-							.addComponent(btnVerwijderen))
-						.addComponent(list, GroupLayout.DEFAULT_SIZE, 236, Short.MAX_VALUE))
+							.addPreferredGap(ComponentPlacement.UNRELATED)
+							.addComponent(btnVerwijderen)
+							.addPreferredGap(ComponentPlacement.UNRELATED)
+							.addComponent(btnPasswordReset, GroupLayout.PREFERRED_SIZE, 29, GroupLayout.PREFERRED_SIZE)))
 					.addContainerGap())
 		);
 		setLayout(groupLayout);
@@ -112,6 +140,23 @@ public class GebruikerBewerkenGui extends JPanel {
 	{
 		this.setVisible(false);
 	}
+	
+	public int OkCancel(String message){
+		int n = JOptionPane.showConfirmDialog(
+                null, message,
+                "Bevestiging",
+                JOptionPane.YES_NO_OPTION);
+		
+		if (n == JOptionPane.YES_OPTION) {
+			return n;
+		} else if (n == JOptionPane.NO_OPTION) {
+			return n;
+		}  
+		return 1;
+		
+	}
+	
+	
 	
 	public Boolean unknownIndex(){
 	if(list.getSelectedValue()==null || list.getSelectedIndex()<0){
@@ -136,9 +181,9 @@ public class GebruikerBewerkenGui extends JPanel {
 					return;
 				} else {
 				
-				System.out.println(list.getSelectedValue().getId());
-				//MedewerkerDAO.removeMedewerker(list.getSelectedIndex());
-				
+					navigation= "gebruikerToevoegen";
+					AdminGui.setHuidigeKeuze(new GebruikerBijwerkenGui(list.getSelectedValue()));
+					
 			}
 				}
 			
@@ -146,14 +191,42 @@ public class GebruikerBewerkenGui extends JPanel {
 				if(!unknownIndex()){
 					return;
 				} else {
+				int n = OkCancel("Ben je zeker dat je " + list.getSelectedValue().getVoornaam() + " " + list.getSelectedValue().getAchternaam() + " wil verwijderen?");	
+					
+				if(n==0){
 				MedewerkerDAO.removeMedewerker(list.getSelectedValue().getId());
 				((DefaultListModel<Medewerker>)list.getModel()).remove(list.getSelectedIndex());
 				JOptionPane.showMessageDialog(new JFrame(), "Gebruiker is succesvol verwijdert.");
+				} else if (n==1){
+					return;
+				}
+				
 			}
+			}
+			
+			if(e.getSource() == btnPasswordReset){
+				if(!unknownIndex()){
+					return;
+				} else {
+					
+					String wachtwoord =  new String("reset1");
+					
+					int n = OkCancel("Ben je zeker dat je een password reset wil uitvoeren op " + list.getSelectedValue().getVoornaam() + " " + list.getSelectedValue().getAchternaam() + "?");
+					
+					if(n==0){
+					LoginDao.updateWachtwoordWhere(list.getSelectedValue().getLogin().getLoginId(), wachtwoord);
+					JOptionPane.showMessageDialog(new JFrame(), "Het wachtwoord van " + list.getSelectedValue().getVoornaam() + " " + list.getSelectedValue().getAchternaam() + " is gereset naar 'reset1'.");
+					} else if (n==1){
+						return;
+					}
+					
+					
+				
+				}
+				
+			
 			}
 			
 		}
 	}
-	
-	
 }
